LoginActivity Done! (without Database)
currently set password and username to 'abc' but the actual thing should be attached to the database to check for the correct username and password. also changed the android manifest so loginactivity is the launcher activity, but still all other activity's parent activity is mainactivty. only mainactivity's parent activity is loginactivity.